presentations/deep-research/demos/pyproject.toml
[project]
name = "deep-research-demos"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.3.0",
    "langchain-tavily>=0.1.0",
    "tavily-python>=0.5.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
